Be 5 add user crud

+++ app/db/models.py
@@ -1,5 +1,6 @@
-from sqlalchemy import Column, Integer, String, Boolean, ARRAY, TIMESTAMP,text
+from sqlalchemy import Column, Integer, String, Boolean, ARRAY, TIMESTAMP,text, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
+
 
 Model = declarative_base()
 
@@ -13,7 +14,7 @@ class User(Model):
     lastname = Column(String)
     status = Column(Boolean, server_default=text('TRUE'))
     city = Column(String)
-    phone = Column(Integer, unique=True)
+    phone = Column(BigInteger, unique=True)
     links = Column(ARRAY(String), server_default=text("'{mylink}'"))
     avatar = Column(String, server_default=text("'myavatar'"))
     hashed_password = Column(String, nullable=False)
